refactor: streamline naming for main page and components
