+++ src/components/main/index.ts
@@ -1,0 +1,5 @@
+import MainBanner from './MainBanner';
+import MainAriticleList from './MainAriticleList';
+import Sidebar from './Sidebar';
+
+export { MainBanner, MainAriticleList, Sidebar };
